refactor: Migrate orbit-core tests to per-module tests/ layout [ORB-00223]
