+++ crates/orbit-core/src/command/job/run/tests/mod.rs
@@ -2,8 +2,8 @@
 
 use crate::OrbitRuntime;
 
-mod cancel;
-mod owner_identity;
+mod actions;
+mod owner;
 mod reconcile;
 
 use chrono::{DateTime, Utc};
