+++ crates/orbit-core/src/config/tests/mod.rs
@@ -1,0 +1,4 @@
+mod agent_detect;
+mod agent_prompt;
+mod bootstrap;
+mod runtime;
